Timing variables for "within 3 months" deprecated and replaced
An error in the logic left a small but significant number of treatment exposures as undeclared (missing) when the screening variable "on_treatment" was marked No but the "hx_treatment" variable was left blank. This error was in der_anytx which propagated to a number of additional second-order and third-order derived variables. A total of 12 variables have been deprecated and replaced with new variables that have addressed the logic error.