Replace der_deadbinary
The original der_deadbinary has been renamed as der_deadbinary_old and deprecated. The original variable did not use the vital status indicators in the baseline and follow-up forms, therefore missing some true deaths. The new variable accounts for some false positives by noting when an indicator is set to deceased on the baseline form, but then subsequent follow-up forms report the patient to be alive (3 = "resuscitated"). All such cases will be centrally adjudicated or queries to the sites to clarify the discrepancies. This change affects a small number of cases but is not felt to be major enough to change dependencies or re-do existing analyses.