Update CCC19_Derived_Variables_Spreadsheet.xlsx
Added an "order of execution" column to capture the order in which variables are created. This is to catch any variables that are depending on derived variables created after the variable. Variables depending on other derived variables that have been verified to fire in the correct order are now highlighted in green in column I